update(pep-8-linting): Linted files, moved level
- Created new module `level-generation`
