+++ requirements.txt
@@ -3,3 +3,4 @@
 # https://pip.pypa.io/en/stable/reference/requirements-file-format/
 
 pygame
+pylint
